added the product form code ProductForm.py
